+++ CHANGELOG.md
@@ -21,3 +21,4 @@
  * Add overloading method for `update` method to update column by multiple condition
  * Log sql info
  * Add id generator strategy for `@Id` field
+ * Add batch insert